refactor: POST request for answers
