+++ server/src/models/Answer.ts
@@ -5,7 +5,6 @@ import { Util } from "./Util";
 export interface Answer extends Util {
   markdown: string;
   questionId: ObjectId;
-  userId: ObjectId;
   upvotes: number;
   downvotes: number;
 }
