+++ server/src/utils/types/AnswerRequestBody.ts
@@ -1,0 +1,4 @@
+export interface AnswerRequestBody {
+  questionId?: string | undefined;
+  markdown?: string | undefined;
+}
