+++ server/src/utils/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./QuestionRequestBody";
+export * from "./AnswerRequestBody";
